Make BP_REST_Activity_Endpoint->create_item() more friendly to BuddyPress plugins
Using the method as a plugin is quite complex as it requires to filter bp_activity_add() making sure a rest request is being performed. I suggest to be nicer with plugins wishing to use the BP Rest API and in particular this activity endpoint by:
Allowing custom components to create items, for instance I see no reason to restrict Activity creation to Core components, it should be available for all active components.
It should allow hidden activities to be created, the Group will need it anywayfor the ! public statuses.

This PR also improves some parts like:
- use a default value for the user ID as bp_loggedin_user_id() is not set in activity functions when creating Group Activities (get_current_user_id() is now used as the default value)
- add the missing unit tests for the create item method when posting group activities.