Add Response headers about total number of activities and pages
Client scripts need to be informed about the total amount of activities found for the request and about the total amount of available pages for the results of the request. Otherwise building a pagination is not possible and it need the script to keep on requesting till the response returns no results.
To avoid altering the response data with these two piece of informations, I suggest to use the allowed headers made available by the WordPress REST API :
- X-WP-Total is informing about the total amount of activities found.
- X-WP-TotalPages is informing about the number of pages.