Activities need to be rendered
Instead of having the raw data fetched into the view context, I think we should provide rendered activities into this context and provide both (raw data + rendered activity) into the edit context.
In other words, we need to make sure the activity content filters are fired to sanitize the output/run formatting filters when the context is a view one.
I suggest to do just like WordPress does for the Post Types : using an object type for the content key of the schema with raw and rendered properties.
From the client side, developers will need to update their scripts to display the content and use data.content.rendered instead of data.content for instance.