Add user favorites
The idea is to implement it a bit differently than how it is done in AJAX. The user is updating their favorites for a specific activity Rest request.
When fetching activities so that it is easy for a client to know whether an activity is favorited or not, i have updated the schema with a readonly property (favorited).
I am only getting the user favorites for get_items/get_item and update_favorite as i dont think we need it elsewhere