Make sure the items proprerty is defined when the parameter type is an array
the `rest_validate_request_arg()` function is checking for the type of the value an array contains. When omitting to set it for array types parameters it generates a notice error. A unit test about requesting group mods is also included as it is the way I found the notice error