Improve the Group Membership Controller
The route to create a new membership should be consistent with what we do in all other enpoints. Thats is to say: do not pass the user ID as the last part of the route. In other words: instead of using `/buddypress/v1/groups/<group_id>/members/<user_id>` we should be using `/buddypress/v1/groups/<group_id>/members` and pass the `user_id` into the request arguments.

This PR brings this improvement. It also adapts the unit tests.
NB: on a site note, the REST Server of unit tests is not setting the current user at the time it fetches the arguments of routes. This means, unlike what happens on the regular WP REST SERVER, the default value set to `bp_loggedin_user_id()` is `0` althought the current user is defined. That is why I added the `user_id` request argument on all `create_item` tests.

I updated the [documentation](https://imath-buddydocs.pf1.wpserveur.net/bp-rest-api/reference/user-groups/group-membership/#add-a-specific-member-into-a-group) to take in account this change.