Invitations: Take current user into account when deleting requests and invitations.
BuddyPress offers a variety of delete functions for invitations and requests, representing different situations.
• `groups_uninvite_user()`
• `groups_reject_invite()`
• `groups_delete_invite()`
• `groups_reject_membership_request()`
• `groups_delete_membership_request()`

I'm assuming that the situation is driven by who is doing the deleting. In this patch, we compare the current user to the subject user of the invitation or request to determine whether the invitation or request is being withdrawn or rejected.