Improve the Group Avatar Controller
Make sure the get_item response is consistent with the item schema. That is to say: always return the full and thumb version of the avatar.
Add the `action` argument to the CREATABLE transport method.
Add a unit test for the `create_item` method.

Write the [documentation](https://imath-buddydocs.pf1.wpserveur.net/bp-rest-api/reference/attachments/group-avatar/) about this controller taking in account these changes.