Improve the item shema and the create_item method
The Signup meta is used by BuddyPress to store the user's full name if the xProdile component is active and most importantly the password the user chose.
I've also added useful information about the date when the activation email was sent and the number of time it was sent to the `get_items()` method for its `edit` context.