Blavatar: simplify the way to get it
It should not be more difficult to get the avatar of the blog from the BP_REST_Attachments_Blog_Avatar_Endpoint than it is from the BP_REST_Blogs_Endpoint. There is no need to ask for the user id, if we use the blog the permission check is using.