Make sure the response of activity posted in a group is the activity
When posting an activity into a private or hidden group, the response does not include the posted activity because in this case `$request[’hidden’]` is not set.
Using `true` for the `show_hidden` argument used into `bp_activity_get` seems fine as anyway the activity has been created by the user, he should be able to read it.