Attachments: Fix uploading for those on Windows/WAMP installs.
As we are using `validate_file()` to validate a full path directory, Windows/WAMP installs should have a full path containing ’:’ as it should starts with ’C:\’. As a result the only thing we need to make sure is avoiding the wrong usage of `../` to validate the full path directory. So if `1` is returned by the `validate_file()` function then we should not allow the upload to carry on.