Members: Only query for fields when necessary.
The WP REST API has a global '_fields' parameter allowing consumers of
the API to restrict the response to only properties specified in
'_fields'.

We should check this parameter before querying for any member
properties. For example, if we add, '_fields=id,name,user_login,link'
to the Members endpoint, this shouldn't query other member fields like
'xprofile' and 'friendship_status'.

This can dramatically speed up the response time for the Members
endpoint when the '_fields' global parameter is used.